+++ srcs/rayTracing/colors.c
@@ -13,7 +13,7 @@ t_rgba separate_color_rgba(int32_t color)
 	return (rgba);
 }
 
-int32_t mix_colors(int32_t colorA, int32_t colorB, double ratio)
+int32_t mix_colors(int32_t colorA, int32_t colorB, float ratio)
 {
 	t_rgba	N;
 	t_rgba	A = separate_color_rgba(colorA);
@@ -22,7 +22,7 @@ int32_t mix_colors(int32_t colorA, int32_t colorB, double ratio)
     // N.r = (int)((1.0 - ratio) * A.r + ratio * B.r);
     // N.g = (int)((1.0 - ratio) * A.g + ratio * B.g);
     // N.b = (int)((1.0 - ratio) * A.b + ratio * B.b);
-(void)ratio;
+	(void)ratio;
 	N.r = clamp((A.r * B.r) / 255,0,255);
 	N.g = clamp((A.g * B.g) / 255,0,255);
 	N.b = clamp((A.b * B.b) / 255,0,255);
@@ -30,7 +30,7 @@ int32_t mix_colors(int32_t colorA, int32_t colorB, double ratio)
     return ((N.r << 24) | (N.g << 16) | (N.b << 8) | N.a);
 }
 
-int32_t brightness(int32_t color, double scale) 
+int32_t brightness(int32_t color, float scale) 
 {
 	t_rgba rgba = separate_color_rgba(color);
 
@@ -41,7 +41,7 @@ int32_t brightness(int32_t color, double scale)
 	return ((rgba.r << 24) | (rgba.g << 16) | (rgba.b << 8) | rgba.a);
 }
 
-double remap(double a, double b, double t)
+float remap(float a, float b, float t)
 {
 	return ((t-a)/(b-a));
 }
